requirements.txt
azure-functions
numerapi
